+++ src/Hash.h
@@ -24,13 +24,13 @@ public:
     Hasher(int bits_per_item, Hash hash_function);
 
     uint32_t hash(std::string item);
-    uint32_t hash(int item);
+    uint32_t hash(uint64_t item);
 
     // same as hash
     uint32_t operator()(std::string item);
 
     uint16_t fingerprint(std::string item);
-    uint16_t fingerprint(size_t item);
+    uint16_t fingerprint(uint64_t item);
 
 };
